requirements-docs.txt
pydata-sphinx-theme>=0.6.3
myst-nb
pymc @ git+https://github.com/pymc-devs/pymc@main
bambi>=0.15.0
sphinx>=4
sphinx-copybutton
sphinx_tabs
sphinx-design
numpydoc
jupyter-sphinx
numpyro>=0.17.0
